+++ packages/spor-react/src/input/SearchInput.tsx
@@ -39,7 +39,7 @@ export const SearchInput = forwardRef<SearchInputProps, "input">(
 
     return (
       <InputGroup position="relative" {...outerProps}>
-        <InputLeftElement>
+        <InputLeftElement pointerEvents="none">
           <SearchOutline24Icon />
         </InputLeftElement>
         <ChakraInput
@@ -48,16 +48,33 @@ export const SearchInput = forwardRef<SearchInputProps, "input">(
           {...innerProps}
           id={inputId}
           type="search"
+          placeholder=" "
           css={{
             "&::-webkit-search-cancel-button": {
               WebkitAppearance: "none",
             },
           }}
           ref={ref}
-          placeholder=" " // This is needed to make the label work as expected
           data-attachable
         />
-        <FormLabel htmlFor={inputId} pointerEvents="none">
+        <FormLabel
+          htmlFor={inputId}
+          sx={{
+            position: "absolute",
+            left: "2.6rem",
+            top: "26.9%",
+            fontSize: "1.13rem",
+            pointerEvents: "none",
+            margin: 0,
+            zIndex: 2,
+            "input:focus + &, input[data-has-value] + &": {
+              color: "var(--chakra-colors-gray-600)",
+            },
+            "input[data-has-value] + &": {
+              transform: "translateY(-40%) scale(0.9)",
+            },
+          }}
+        >
           {label ?? t(texts.label)}
         </FormLabel>
         {showClearButton && (