+++ packages/spor-react/src/theme/components/textarea.ts
@@ -4,13 +4,13 @@ import { inputBaseStyle, inputVariant } from "../utils/input-utils";
 const config = defineStyleConfig({
   baseStyle: (props) => ({
     ...inputBaseStyle(props).field,
-    minHeight: "5rem",
+    minHeight: "calc(var(--label-height) + 4rem)",
     verticalAlign: "top",
     appearance: "none",
-    paddingTop: 2,
+    borderTop: "0.8rem solid transparent",
     "&:not(:placeholder-shown)": {
       "&:has(+ label)": {
-        paddingTop: 4,
+        borderTop: "var(--label-height) solid transparent", // use border in stead of padding to avoid problems with scrolling
       },
       "& + label": {
         transform: "scale(0.825) translateY(-10px)",